+++ src/main/resources/db/changelog/sql/01.000.00/4.sql
@@ -2,4 +2,4 @@
 
 --changeset draft:1
 ALTER TABLE strategy
-    ADD COLUMN minLot INTEGER;
+    ADD COLUMN minLot numeric(20,5);
