+++ src/main/resources/db/changelog/sql/01.000.00/5.sql
@@ -3,10 +3,10 @@
 --changeset draft:1
 CREATE TABLE LAST_PRICE
 (
-    "id"             SERIAL unique not null,
-    "figi"           varchar(255),
+    "id"              SERIAL unique not null,
+    "figi"            varchar(255),
     "name_subscriber" varchar(20),
-    "price"          float,
+    "price"           numeric(20, 5),
     "update_time"     TIMESTAMP,
     CONSTRAINT "LastPrice_pk" PRIMARY KEY ("id")
 );
